Ensure there is at least the default edition in the query QueryParams to a tools package
